feat: added certification supported document upload option for doctor and hospital (DBIO-418 & DBIO-419).
