+++ src/views/Dashboard/Hospital/Account/index.vue
@@ -118,130 +118,26 @@
               </v-form>
             </v-card-text>
           </v-card>
-
-          <v-card class="dg-card mt-5" elevation="0" outlined>
-            <v-card-text class="px-8 mt-5">
-              <div class="d-flex justify-space-between align-center">
-                <div class="secondary--text text-h6">
-                  <b>Certifications</b>
-                </div>
-                <v-btn small dark color="#75DEE4" fab style="border-radius:10px;" @click="certificationDialog = true">
-                  <v-icon>mdi-plus</v-icon>
-                </v-btn>
-              </div>
-              <div v-if="hospitalAccount.certifications.length == 0">
-                You don’t have any certifications
-              </div>
-              <div v-if="isLoading" class="mt-5">
-                <v-skeleton-loader 
-                  v-for="data in hospitalAccount.certifications"
-                  :key="data.idx"
-                  type="list-item-three-line"
-                  min-width="200"
-                ></v-skeleton-loader>
-              </div>
-              <div v-if="hospitalAccount.certifications.length > 0 && !isLoading" class="mt-5">
-                <div
-                  v-for="(cert, idx) in hospitalAccount.certifications"
-                  :key="cert.id"
-                  :style="idx < (hospitalAccount.certifications.length - 1) && 'border-bottom: 1px solid #555454;'"
-                  class="my-3"
-                >
-                  <div class="d-flex justify-space-between align-center" style="width: 100%;">
-                    <div class=""><b>{{ cert.info.title }}</b></div>
-                    <div class="d-flex">
-                      <v-icon class="mx-1" small @click="editCertification(cert)">mdi-pencil</v-icon>
-                      <v-icon class="mx-1" small @click="deleteCertification(cert)">mdi-delete</v-icon>
-                    </div>
-                  </div>
-                  <div>{{ cert.info.month }} {{ cert.info.year }} • {{ cert.info.issuer }}</div>
-                  <div class="mt-3 mb-3">{{ cert.info.description }}</div>
-                </div>
-              </div>
-            </v-card-text>
-          </v-card>
+          
+          <HospitalCertification />
         </v-col>
       </v-row>
     </v-container>
-
-    <Dialog :show="certificationDialog" @close="closeCertificationDialog">
-      <template v-slot:title>
-        <div class="secondary--text h6">
-          {{ isEditCertificationDialog ? "Edit" : "Add" }} Certification
-        </div>
-      </template>
-      <template v-slot:body>
-        <v-form ref="certificationForm">
-          <v-text-field
-            dense
-            label="Title"
-            placeholder="Title"
-            outlined
-            v-model="certTitle"
-            :rules="[val => !!val || 'Title is required']"
-            ></v-text-field>
-          <v-text-field
-            dense
-            label="Issuer"
-            placeholder="Issuer"
-            outlined
-            v-model="certIssuer"
-            :rules="[val => !!val || 'Issuer is required']"
-            ></v-text-field>
-          <div class="d-flex justify-space-between align-center">
-            <div class="mr-1">
-              <v-select
-                dense
-                label="Month"
-                :items="selectMonths"
-                outlined
-                v-model="certMonth"
-                :rules="[val => !!val || 'Month is required']"
-                ></v-select>
-            </div>
-            <div class="ml-1">
-              <v-select
-                dense
-                label="Year"
-                :items="selectYears"
-                outlined
-                v-model="certYear"
-                :rules="[val => !!val || 'Year is required']"
-                ></v-select>
-            </div>
-          </div>
-          <v-textarea
-            outlined
-            label="Description"
-            v-model="certDescription"
-            :rules="[val => !!val || 'Description is required']"
-          ></v-textarea>
-        </v-form>
-      </template>
-      <template v-slot:actions>
-        <Button @click="submitCertification" :loading="isLoading" color="primary" dark>
-          Save
-        </Button>
-      </template>
-    </Dialog>
   </div>
 </template>
 
 <script>
 import { mapGetters } from 'vuex'
 import { updateHospital } from '@/lib/polkadotProvider/command/hospitals'
-import { createCertification, updateCertification, deleteCertification } from '@/lib/polkadotProvider/command/hospitals/certifications'
 import countryData from "@/assets/json/country.json"
 import cityData from "@/assets/json/city.json"
 import { upload } from "@/lib/ipfs"
-import Dialog from '@/components/Dialog'
-import Button from '@/components/Button'
+import HospitalCertification from "./Certification"
 
 export default {
   name: 'HospitalAccount',
   components: {
-    Dialog,
-    Button,
+    HospitalCertification,
   },
   async mounted() {
     const hospitalInfo = this.hospitalAccount.info
@@ -275,18 +171,8 @@ export default {
     cities: [],
     image: "",
     files: [],
-    certifications: [],
-    certificationDialog: false,
-    certId: "", // for update certification
-    certTitle: "",
-    certIssuer: "",
-    certMonth: "",
-    certYear: "",
-    certDescription: "",
-    selectMonths: ["January", "February", "March", "April", "May", "June", "July", "August", "September", "October", "November", "December"],
     isLoading: false,
     isUploading: false,
-    isEditCertificationDialog: false,
   }),
   computed: {
     ...mapGetters({
@@ -294,19 +180,6 @@ export default {
       pair: 'substrate/wallet',
       hospitalAccount: 'substrate/hospitalAccount',
     }),
-    citiesSelection() {
-      return this.cities
-        .filter((c) => c.country == this.country)
-        .map((c) => ({ value: c.city, text: c.city, country: c.country }));
-    },
-    selectYears() {
-      const years = []
-      const thisYear = new Date().getFullYear()
-      for (let i = thisYear; i >= 2000; i--) {
-        years.push(String(i))
-      }
-      return years
-    },
   },
   methods: {
     async getCountries() {
@@ -373,89 +246,6 @@ export default {
         this.files = []
         this.image = ''
       }
-    },
-    closeCertificationDialog() {
-      this.certId = ""
-      this.certificationDialog = false
-      this.isEditCertificationDialog = false
-      this.$refs.certificationForm.reset()
-    },
-    async submitCertification() {
-      if (!this.certId) {
-        await this.addCertification()
-        return
-      }
-      await this.updateCertification()
-    },
-    async addCertification() {
-      if (!this.$refs.certificationForm.validate()) {
-        return
-      }
-      try {
-        this.isLoading = true
-        const certificationInfo = {
-          title: this.certTitle,
-          issuer: this.certIssuer,
-          month: this.certMonth,
-          year: this.certYear,
-          description: this.certDescription,
-        }
-
-        await createCertification(this.api, this.pair, certificationInfo, () => {
-          this.closeCertificationDialog()
-          this.isLoading = false
-        })
-      } catch (err) {
-        console.log(err)
-        this.isLoading = false
-      }
-    },
-    editCertification(cert) {
-      this.certId = cert.id
-      this.certTitle = cert.info.title
-      this.certIssuer = cert.info.issuer
-      this.certMonth = cert.info.month
-      this.certYear = cert.info.year
-      this.certDescription = cert.info.description
-
-      this.certificationDialog = true
-      this.isEditCertificationDialog = true
-    },
-    async updateCertification() {
-      if (!this.$refs.certificationForm.validate()) {
-        return
-      }
-
-      try {
-        this.isLoading = true
-
-        const certificationInfo = {
-          title: this.certTitle,
-          issuer: this.certIssuer,
-          month: this.certMonth,
-          year: this.certYear,
-          description: this.certDescription,
-        }
-
-        await updateCertification(this.api, this.pair, this.certId, certificationInfo, () => {
-          this.closeCertificationDialog()
-          this.isLoading = false
-        })
-      } catch (err) {
-        console.log(err)
-        this.isLoading = false
-      }
-    },
-    async deleteCertification(cert) {
-      const isConfirmed = confirm("Are you sure you want to delete this certification?")
-      if (isConfirmed) {
-        this.isLoading = true
-
-        await deleteCertification(this.api, this.pair, cert.id, () => {
-          this.isLoading = false
-        })
-      }
-      return
     },
   }
 }
